+++ src/main/java/com/github/antag99/retinazer/EntitySystem.java
@@ -7,4 +7,23 @@ public abstract class EntitySystem implements EngineService {
     @Wire
     protected Engine engine;
 
+    private final Family family;
+    private EntitySetView familyEntities;
+
+    protected EntitySystem(Family family) {
+        this.family = family;
+    }
+
+    public final Family getFamily() {
+        return family;
+    }
+
+    public final EntitySetView getEntities() {
+        return familyEntities;
+    }
+
+    @Override
+    public void initialize() {
+        familyEntities = engine.getEntities(family);
+    }
 }